requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
playwright-stealth==1.0.6
python-dotenv==1.0.0
pydantic==2.4.2
PyYAML==6.0.1
aiosqlite==0.19.0
sse-starlette==1.8.2
